Rename MyApp to FlapApp
