+++ lib/presentation/flap_app.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-class MyApp extends ConsumerWidget {
-  const MyApp({super.key});
+class FlapApp extends ConsumerWidget {
+  const FlapApp({super.key});
 
   // This widget is the root of your application.
   @override
